refactor(tests): extract inline test modules to tests/ directories (#2462-#2466)
Move #[cfg(test)] modules from lib.rs to separate tests/ files for
perl-lsp-perltidy, perl-lsp-document-highlight, perl-lsp-on-type-formatting,
perl-dap-config, and perl-pod. Private helper tests that access non-pub
functions stay in lib.rs (on-type-formatting has 7); all public-API tests
move to integration test files. 148 tests total, all passing.
