+++ crates/perl-lsp-on-type-formatting/Cargo.toml
@@ -14,6 +14,7 @@ keywords = ["perl", "lsp", "formatting", "on-type", "ide"]
 categories = ["development-tools", "text-editors"]
 include = [
   "src/**",
+  "tests/**",
   "Cargo.toml",
   "README.md",
   "LICENSE*"
