+++ crates/perl-lsp-on-type-formatting/src/lib.rs
@@ -325,267 +325,7 @@ fn find_heredoc_tags(line: &str) -> Vec<String> {
 mod tests {
     use super::*;
 
-    // ------------------------------------------------------------------
-    // Helper: extract the "newText" from the first edit in the result.
-    // ------------------------------------------------------------------
-    fn first_new_text(edits: &Option<Vec<Value>>) -> Option<String> {
-        edits
-            .as_ref()
-            .and_then(|v| v.first())
-            .and_then(|e| e.get("newText"))
-            .and_then(|t| t.as_str())
-            .map(String::from)
-    }
-
-    fn first_edit_range(edits: &Option<Vec<Value>>) -> Option<(u64, u64, u64, u64)> {
-        edits.as_ref().and_then(|v| v.first()).and_then(|e| {
-            let r = e.get("range")?;
-            let sl = r.pointer("/start/line")?.as_u64()?;
-            let sc = r.pointer("/start/character")?.as_u64()?;
-            let el = r.pointer("/end/line")?.as_u64()?;
-            let ec = r.pointer("/end/character")?.as_u64()?;
-            Some((sl, sc, el, ec))
-        })
-    }
-
-    // ==================================================================
-    //  `}` — auto-indent to match opening `{`
-    // ==================================================================
-
-    #[test]
-    fn close_brace_aligns_to_opening_brace() {
-        // The `}` on line 2 is indented 4 spaces, but the opening `{` is
-        // on a line with 0-space indent, so `}` should be re-indented to 0.
-        let text = "if ($x) {\n  print 1;\n    }";
-        let edits = compute_on_type_edit(text, 2, 5, '}');
-        assert!(edits.is_some());
-        let new_text = first_new_text(&edits);
-        assert_eq!(new_text.as_deref(), Some(""));
-        // Range should replace the leading 4 spaces.
-        assert_eq!(first_edit_range(&edits), Some((2, 0, 2, 4)));
-    }
-
-    #[test]
-    fn close_brace_already_correct_indentation() {
-        let text = "if ($x) {\n  print 1;\n}";
-        let edits = compute_on_type_edit(text, 2, 1, '}');
-        // Already at correct indent (0), so no edits.
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn close_brace_nested_blocks() {
-        let text = "sub foo {\n  if ($x) {\n    print 1;\n      }\n}";
-        // Line 3: "      }" has 6 spaces, should align to "  if ($x) {" = 2 spaces.
-        let edits = compute_on_type_edit(text, 3, 7, '}');
-        assert!(edits.is_some());
-        assert_eq!(first_new_text(&edits).as_deref(), Some("  "));
-        assert_eq!(first_edit_range(&edits), Some((3, 0, 3, 6)));
-    }
-
-    #[test]
-    fn close_brace_on_first_line_uses_saturating_sub() {
-        // `}` on line 0 with no matching opener — falls back to saturating_sub.
-        let text = "  }";
-        let edits = compute_on_type_edit(text, 0, 3, '}');
-        assert!(edits.is_some());
-        assert_eq!(first_new_text(&edits).as_deref(), Some(""));
-    }
-
-    #[test]
-    fn close_brace_skips_braces_in_strings() {
-        // The `{` inside the string should not count as the matching opener.
-        let text = "my $x = \"{\";\nif ($y) {\n  1;\n    }";
-        let edits = compute_on_type_edit(text, 3, 5, '}');
-        assert!(edits.is_some());
-        // Should align to line 1 ("if ($y) {") which has indent 0.
-        assert_eq!(first_new_text(&edits).as_deref(), Some(""));
-    }
-
-    #[test]
-    fn close_brace_skips_braces_in_comments() {
-        // The `{` in the comment should be ignored.
-        let text = "# {\nif ($y) {\n  1;\n    }";
-        let edits = compute_on_type_edit(text, 3, 5, '}');
-        assert!(edits.is_some());
-        assert_eq!(first_new_text(&edits).as_deref(), Some(""));
-    }
-
-    // ==================================================================
-    //  `;` — does not change indentation
-    // ==================================================================
-
-    #[test]
-    fn semicolon_does_not_change_indentation() {
-        let text = "  my $x = 1;";
-        let edits = compute_on_type_edit(text, 0, 12, ';');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn semicolon_at_end_of_indented_line_no_edit() {
-        let text = "sub foo {\n    my $x = 1;";
-        let edits = compute_on_type_edit(text, 1, 14, ';');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn semicolon_preserves_deeper_indentation() {
-        let text = "        return 1;";
-        let edits = compute_on_type_edit(text, 0, 17, ';');
-        assert!(edits.is_none());
-    }
-
-    // ==================================================================
-    //  `\n` — indent new line after `{`
-    // ==================================================================
-
-    #[test]
-    fn newline_after_open_brace_indents() {
-        let text = "if ($x) {\n";
-        let edits = compute_on_type_edit(text, 1, 0, '\n');
-        assert!(edits.is_some());
-        // New line should get 2-space indent (0 + 2).
-        assert_eq!(first_new_text(&edits).as_deref(), Some("  "));
-    }
-
-    #[test]
-    fn newline_after_indented_open_brace() {
-        let text = "  sub foo {\n";
-        let edits = compute_on_type_edit(text, 1, 0, '\n');
-        assert!(edits.is_some());
-        // Previous line has 2-space indent, so new line gets 4.
-        assert_eq!(first_new_text(&edits).as_deref(), Some("    "));
-    }
-
-    #[test]
-    fn newline_after_plain_statement_keeps_indent() {
-        let text = "    my $x = 1;\n";
-        let edits = compute_on_type_edit(text, 1, 0, '\n');
-        assert!(edits.is_some());
-        assert_eq!(first_new_text(&edits).as_deref(), Some("    "));
-    }
-
-    #[test]
-    fn newline_after_close_brace_keeps_its_indent() {
-        let text = "  if ($x) {\n    1;\n  }\n";
-        // Line 3 is after "  }", which has 2-space indent.
-        let edits = compute_on_type_edit(text, 3, 0, '\n');
-        assert!(edits.is_some());
-        assert_eq!(first_new_text(&edits).as_deref(), Some("  "));
-    }
-
-    #[test]
-    fn newline_on_first_line_returns_none() {
-        let text = "\n";
-        let edits = compute_on_type_edit(text, 0, 0, '\n');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn newline_replaces_existing_wrong_indent() {
-        let text = "if ($x) {\n      ";
-        // Line 1 has 6-space indent but should be 2.
-        let edits = compute_on_type_edit(text, 1, 6, '\n');
-        assert!(edits.is_some());
-        assert_eq!(first_new_text(&edits).as_deref(), Some("  "));
-        assert_eq!(first_edit_range(&edits), Some((1, 0, 1, 6)));
-    }
-
-    // ==================================================================
-    //  Heredoc — formatting is disabled
-    // ==================================================================
-
-    #[test]
-    fn heredoc_suppresses_close_brace() {
-        let text = "my $x = <<END;\n}\nEND";
-        let edits = compute_on_type_edit(text, 1, 1, '}');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn heredoc_suppresses_semicolon() {
-        let text = "my $x = <<END;\nsome text;\nEND";
-        let edits = compute_on_type_edit(text, 1, 10, ';');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn heredoc_suppresses_newline() {
-        let text = "my $x = <<END;\nsome text\n\nEND";
-        let edits = compute_on_type_edit(text, 2, 0, '\n');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn heredoc_single_quoted_tag() {
-        let text = "my $x = <<'END';\n}\nEND";
-        let edits = compute_on_type_edit(text, 1, 1, '}');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn heredoc_double_quoted_tag() {
-        let text = "my $x = <<\"END\";\n}\nEND";
-        let edits = compute_on_type_edit(text, 1, 1, '}');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn heredoc_indented_form() {
-        let text = "my $x = <<~END;\n  }\n  END";
-        let edits = compute_on_type_edit(text, 1, 3, '}');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn after_heredoc_terminator_formatting_resumes() {
-        let text = "my $x = <<END;\nheredoc body\nEND\n    }";
-        // Line 3 ("    }") is after the heredoc ends, so formatting should work.
-        let edits = compute_on_type_edit(text, 3, 5, '}');
-        // No matching opener, so falls back to saturating_sub: 4 - 2 = 2.
-        assert!(edits.is_some());
-        assert_eq!(first_new_text(&edits).as_deref(), Some("  "));
-    }
-
-    // ==================================================================
-    //  Edge cases
-    // ==================================================================
-
-    #[test]
-    fn out_of_bounds_line_is_none() {
-        let edits = compute_on_type_edit("", 42, 0, '}');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn returns_none_for_unknown_trigger() {
-        let text = "my $x = 1;";
-        let edits = compute_on_type_edit(text, 0, 5, 'a');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn empty_document_newline() {
-        let edits = compute_on_type_edit("\n", 0, 0, '\n');
-        assert!(edits.is_none());
-    }
-
-    #[test]
-    fn multiple_heredocs_on_one_line() {
-        // Two heredocs on one line: body of first, then body of second.
-        let text = "my ($a, $b) = (<<A, <<B);\nfirst\nA\nsecond\nB\nnormal;";
-        // Line 1 ("first") is inside heredoc A.
-        assert!(compute_on_type_edit(text, 1, 5, ';').is_none());
-        // Line 3 ("second") is inside heredoc B.
-        assert!(compute_on_type_edit(text, 3, 6, ';').is_none());
-        // Line 5 ("normal;") is after both heredocs.
-        assert!(compute_on_type_edit(text, 5, 7, ';').is_none()); // `;` is always None
-    }
-
-    // ------------------------------------------------------------------
-    //  Internal helper unit tests
-    // ------------------------------------------------------------------
+    // Internal helper unit tests — these call private functions and must stay here.
 
     #[test]
     fn extract_braces_skips_strings_and_comments() {
