+++ crates/perl-pod/src/lib.rs
@@ -289,370 +289,3 @@ fn extract_link_display(link: &str) -> String {
 fn is_pod_format_code(c: char) -> bool {
     matches!(c, 'B' | 'I' | 'C' | 'L' | 'F' | 'S' | 'E' | 'X' | 'Z')
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn empty_source_returns_empty_doc() {
-        let doc = extract_pod("");
-        assert!(doc.is_empty());
-    }
-
-    #[test]
-    fn pure_code_no_pod() {
-        let source = r#"
-package Foo::Bar;
-use strict;
-sub new { bless {}, shift }
-1;
-"#;
-        let doc = extract_pod(source);
-        assert!(doc.is_empty());
-    }
-
-    #[test]
-    fn extracts_name_section() {
-        let source = r#"
-=head1 NAME
-
-Foo::Bar - A sample module
-
-=cut
-"#;
-        let doc = extract_pod(source);
-        assert_eq!(doc.name.as_deref(), Some("Foo::Bar - A sample module"));
-    }
-
-    #[test]
-    fn extracts_synopsis() {
-        let source = r#"
-=head1 SYNOPSIS
-
-    use Foo::Bar;
-    my $obj = Foo::Bar->new();
-
-=cut
-"#;
-        let doc = extract_pod(source);
-        assert!(doc.synopsis.is_some());
-        assert!(doc.synopsis.as_ref().is_some_and(|s| s.contains("use Foo::Bar")));
-    }
-
-    #[test]
-    fn extracts_description_first_paragraph() {
-        let source = r#"
-=head1 DESCRIPTION
-
-This module does amazing things.
-It is very useful.
-
-This second paragraph should not be included.
-
-=cut
-"#;
-        let doc = extract_pod(source);
-        assert_eq!(
-            doc.description.as_deref(),
-            Some("This module does amazing things.\nIt is very useful.")
-        );
-    }
-
-    #[test]
-    fn extracts_methods() {
-        let source = r#"
-=head2 new
-
-Creates a new instance of the object.
-
-=head2 process
-
-Processes the input data.
-
-=cut
-"#;
-        let doc = extract_pod(source);
-        assert_eq!(doc.methods.len(), 2);
-        assert!(doc.methods.contains_key("new"));
-        assert!(doc.methods.contains_key("process"));
-        assert!(doc.methods["new"].contains("Creates a new instance"));
-        assert!(doc.methods["process"].contains("Processes the input data"));
-    }
-
-    #[test]
-    fn strips_bold_formatting() {
-        assert_eq!(strip_pod_formatting("B<bold text>"), "bold text");
-    }
-
-    #[test]
-    fn strips_italic_formatting() {
-        assert_eq!(strip_pod_formatting("I<italic text>"), "italic text");
-    }
-
-    #[test]
-    fn strips_code_formatting() {
-        assert_eq!(strip_pod_formatting("C<my $var>"), "my $var");
-    }
-
-    #[test]
-    fn strips_link_simple() {
-        assert_eq!(strip_pod_formatting("L<Module::Name>"), "Module::Name");
-    }
-
-    #[test]
-    fn strips_link_with_display_text() {
-        assert_eq!(strip_pod_formatting("L<click here|Module::Name>"), "click here");
-    }
-
-    #[test]
-    fn strips_link_with_section() {
-        assert_eq!(strip_pod_formatting("L<Module::Name/method>"), "Module::Name");
-    }
-
-    #[test]
-    fn mixed_formatting() {
-        assert_eq!(
-            strip_pod_formatting("Use B<new> to create a C<Foo> object"),
-            "Use new to create a Foo object"
-        );
-    }
-
-    #[test]
-    fn handles_cut_properly() {
-        let source = r#"
-=head1 NAME
-
-First - Module
-
-=cut
-
-package First;
-
-=head1 NAME
-
-Second - Module
-
-=cut
-"#;
-        let doc = extract_pod(source);
-        // Second =head1 NAME overwrites the first
-        assert_eq!(doc.name.as_deref(), Some("Second - Module"));
-    }
-
-    #[test]
-    fn handles_pod_without_cut_at_eof() {
-        let source = r#"
-=head1 NAME
-
-Foo::Bar - No cut at end
-"#;
-        let doc = extract_pod(source);
-        assert_eq!(doc.name.as_deref(), Some("Foo::Bar - No cut at end"));
-    }
-
-    #[test]
-    fn handles_over_item_back() {
-        let source = r#"
-=head2 options
-
-Available options:
-
-=over 4
-
-=item B<verbose>
-
-Enable verbose output.
-
-=item B<quiet>
-
-Suppress output.
-
-=back
-
-=cut
-"#;
-        let doc = extract_pod(source);
-        assert!(doc.methods.contains_key("options"));
-        let method_doc = &doc.methods["options"];
-        assert!(method_doc.contains("Available options:"));
-        assert!(method_doc.contains("- verbose"));
-        assert!(method_doc.contains("- quiet"));
-    }
-
-    #[test]
-    fn full_module_extraction() {
-        let source = r#"
-package DateTime::Format::Custom;
-
-use strict;
-use warnings;
-
-=head1 NAME
-
-DateTime::Format::Custom - Parse and format dates
-
-=head1 SYNOPSIS
-
-    use DateTime::Format::Custom;
-    my $dt = DateTime::Format::Custom->parse("2024-01-01");
-
-=head1 DESCRIPTION
-
-This module provides custom date parsing and formatting
-capabilities for the DateTime ecosystem.
-
-It supports multiple input formats and can auto-detect
-the format of input strings.
-
-=head2 parse
-
-    my $dt = DateTime::Format::Custom->parse($string);
-
-Parses a date string and returns a L<DateTime> object.
-
-=head2 format
-
-    my $str = DateTime::Format::Custom->format($dt);
-
-Formats a B<DateTime> object as a string.
-
-=head1 AUTHOR
-
-Jane Doe
-
-=cut
-
-sub parse { ... }
-sub format { ... }
-
-1;
-"#;
-        let doc = extract_pod(source);
-        assert_eq!(doc.name.as_deref(), Some("DateTime::Format::Custom - Parse and format dates"));
-        assert!(doc.synopsis.as_ref().is_some_and(|s| s.contains("use DateTime::Format::Custom")));
-        assert!(doc.description.as_ref().is_some_and(|s| s.contains("custom date parsing")));
-        // Description should only be first paragraph
-        assert!(!doc.description.as_ref().is_none_or(|s| s.contains("auto-detect")));
-        assert_eq!(doc.methods.len(), 2);
-        assert!(doc.methods["parse"].contains("Parses a date string"));
-        assert!(doc.methods["parse"].contains("DateTime"));
-        assert!(doc.methods["format"].contains("Formats a DateTime object"));
-    }
-
-    #[test]
-    fn extract_pod_from_file_missing_file() {
-        let result = extract_pod_from_file(Path::new("/nonexistent/file.pm"));
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn nested_formatting_codes() {
-        // Depth tracking handles nested angle brackets correctly
-        assert_eq!(strip_pod_formatting("B<I<bold italic>>"), "bold italic");
-    }
-
-    #[test]
-    fn no_formatting_passthrough() {
-        assert_eq!(strip_pod_formatting("plain text here"), "plain text here");
-    }
-
-    #[test]
-    fn empty_formatting_code() {
-        assert_eq!(strip_pod_formatting("B<>"), "");
-        assert_eq!(strip_pod_formatting("C<>"), "");
-    }
-
-    #[test]
-    fn head1_other_sections_ignored() {
-        let source = r#"
-=head1 AUTHOR
-
-John Doe
-
-=head1 LICENSE
-
-Same as Perl itself.
-
-=cut
-"#;
-        let doc = extract_pod(source);
-        assert!(doc.name.is_none());
-        assert!(doc.synopsis.is_none());
-        assert!(doc.description.is_none());
-        assert!(doc.methods.is_empty());
-    }
-
-    #[test]
-    fn pod_directive_starts_pod_mode() {
-        let source = r#"
-package Foo;
-
-=pod
-
-This is some POD text.
-
-=head1 NAME
-
-Foo - A module
-
-=cut
-"#;
-        let doc = extract_pod(source);
-        assert_eq!(doc.name.as_deref(), Some("Foo - A module"));
-    }
-
-    #[test]
-    fn encoding_directive_starts_pod() {
-        let source = r#"
-=encoding utf-8
-
-=head1 NAME
-
-Encoded::Module - Uses UTF-8
-
-=cut
-"#;
-        let doc = extract_pod(source);
-        assert_eq!(doc.name.as_deref(), Some("Encoded::Module - Uses UTF-8"));
-    }
-
-    #[test]
-    fn f_format_code_for_filenames() {
-        assert_eq!(strip_pod_formatting("See F<config.yml>"), "See config.yml");
-    }
-
-    #[test]
-    fn e_format_code_passthrough() {
-        // E<lt> E<gt> etc. — we just strip the code, leaving the entity name
-        assert_eq!(strip_pod_formatting("E<lt>"), "lt");
-    }
-
-    #[test]
-    fn multiple_pod_blocks() {
-        let source = r#"
-package Multi;
-
-=head1 NAME
-
-Multi - Multiple POD blocks
-
-=cut
-
-sub helper { 1 }
-
-=head2 run
-
-Runs the main logic.
-
-=cut
-
-sub run { }
-
-1;
-"#;
-        let doc = extract_pod(source);
-        assert_eq!(doc.name.as_deref(), Some("Multi - Multiple POD blocks"));
-        assert!(doc.methods.contains_key("run"));
-    }
-}
